+++ src/components/IdeaListItem.vue
@@ -35,7 +35,7 @@
       <q-item-tile sublabel lines="1" v-if="showAuthor">{{ idea.author }}</q-item-tile>
     </q-item-side>
     <q-modal v-model="editOpen" content-classes="edit-idea-modal">
-      <q-icon name="fa-close" size="2rem" @click="editOpen=false" class="float-right cursor-pointer" />
+      <q-icon name="fa-close" size="2rem" @click.native="editOpen=false" class="float-right cursor-pointer" />
       <idea-form ref="ideaForm" :idea="idea" :modal-fin="closeEdit" form-type="Edit" />
     </q-modal>
   </q-item>