+++ src/components/IllustrationListItem.vue
@@ -36,7 +36,7 @@
       <q-item-tile sublabel lines="1" v-if="showAuthor">{{ illustration.author }}</q-item-tile>
     </q-item-side>
     <q-modal v-model="editOpen" content-classes="edit-illustration-modal">
-      <q-icon name="fa-close" size="2rem" @click="editOpen=false" class="float-right cursor-pointer" />
+      <q-icon name="fa-close" size="2rem" @click.native="editOpen=false" class="float-right cursor-pointer" />
       <illustration-form ref="illustrationForm" :illustration="illustration" :modal-fin="closeEdit" form-type="Edit" />
     </q-modal>
   </q-item>